fix: separate profile edit input requirements

+++ app/Http/Controllers/ProfileController.php
@@ -24,33 +24,39 @@ class ProfileController extends Controller
 	{
 		$user = Auth::user();
 
-		if ($user->profile_image) {
-			Storage::disk('public')->delete($user->profile_image);
+		if ($request->update_profile_image === '1') {
+			if ($user->profile_image) {
+				Storage::disk('public')->delete($user->profile_image);
 
-			$user->profile_image = null;
+				$user->profile_image = null;
+			}
+
+			if ($request->hasFile('profile_image')) {
+				$path = $request->file('profile_image')->store('profile_images', 'public');
+
+				$user->profile_image = $path;
+			}
 		}
 
-		if ($request->hasFile('profile_image')) {
-			$path = $request->file('profile_image')->store('profile_images', 'public');
+		if ($request->update_cover_image === '1') {
+			$coverImg = Settings::first();
 
-			$user->profile_image = $path;
+			if ($coverImg) {
+				Storage::disk('public')->delete($coverImg->cover_image);
+
+				$coverImg->delete();
+			}
+
+			if ($request->hasFile('cover_image')) {
+				$path = $request->file('cover_image')->store('cover_images', 'public');
+
+				Settings::create(['cover_image' => $path]);
+			}
 		}
 
-		$coverImg = Settings::first();
-
-		if ($coverImg) {
-			Storage::disk('public')->delete($coverImg->cover_image);
-
-			$coverImg->delete();
+		if ($request->filled('password')) {
+			$user->password = $request->password;
 		}
-
-		if ($request->hasFile('cover_image')) {
-			$path = $request->file('cover_image')->store('cover_images', 'public');
-
-			Settings::create(['cover_image' => $path]);
-		}
-
-		$user->password = $request->password;
 
 		$user->save();
